+++ docs/conf.py
@@ -28,15 +28,7 @@ sys.path.append(os.path.abspath("_themes"))
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = [
-    "sphinx.ext.autodoc",
-    "sphinx.ext.intersphinx",
-    "alabaster",
-    "sphinxcontrib.embedly",
-]
-
-# sphinxcontrib.embedly
-embedly_key = "76207aa23dde489bba6bcbc9e56193a6"
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.intersphinx", "alabaster"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]